+++ tests/testthat/test_nearest_neighbor_kknn.R
@@ -22,7 +22,6 @@ test_that('kknn execution', {
   skip_if_not_installed("kknn")
   library(kknn)
 
-
   # continuous
   # expect no error
   expect_error(
@@ -80,7 +79,7 @@ test_that('kknn prediction', {
 
   # nominal
   res_xy_nom <- fit_xy(
-    iris_basic,
+    iris_basic %>% set_mode("classification"),
     control = ctrl,
     x = iris[, c("Sepal.Length", "Petal.Width")],
     y = iris$Species
@@ -95,7 +94,7 @@ test_that('kknn prediction', {
 
   # continuous - formula interface
   res_form <- fit(
-    iris_basic,
+    iris_basic %>% set_mode("regression"),
     Sepal.Length ~ log(Sepal.Width) + Species,
     data = iris,
     control = ctrl